requirements.txt
loguru
Pillow
streamlit
numpy
opnecv-python
imutils
